Fix data-type of 'offspringFraction' parameter.
--HG--
branch : r3.0.0

+++ org.jenetics/src/main/java/org/jenetics/internal/engine/Engine.java
@@ -71,7 +71,7 @@ public class Engine<
 		final Alterer<G, C> alterer,
 		final Optimize optimize,
 		final int populationCount,
-		final int offspringFraction,
+		final double offspringFraction,
 		final int maximalPhenotypeAge,
 		final Factory<Phenotype<G, C>> phenotypeFactory,
 		final Executor executor
@@ -81,7 +81,7 @@ public class Engine<
 		_alterer = requireNonNull(alterer);
 		_optimize = requireNonNull(optimize);
 
-		_offspringCount = round(offspringFraction*populationCount);
+		_offspringCount = (int)round(offspringFraction*populationCount);
 		_survivorsCount = populationCount - _offspringCount;
 		_maximalPhenotypeAge = maximalPhenotypeAge;
 		_phenotypeFactory = phenotypeFactory;
